[Gold V] Title: 동전 1, Time: 0 ms, Memory: 2292 KB -BaekjoonHub

+++ 백준/Gold/2293. 동전 1/동전 1.cc
@@ -5,12 +5,14 @@ int main(){
     int n;
     int k;
     int arr[101];
-    long long dp [100001];
+    int dp [100001];
+
     cin >> n >> k;
 
     for(int i =0 ; i< n ; i++){
         cin >> arr[i];
     }
+
     dp[0]=1;
 
     for(int i = 0;i<n;i++){
@@ -21,8 +23,6 @@ int main(){
 
     cout << dp[k] << endl;
     return 0;
-
-
